Ignoring Unity generated .sln and .csproj
Deleting the unity .sln

Deleting the unity .csproj

Fixed GreyGuy pngs to be Sprites again

Broke the GreyGuy files when cleaning up meta tags, now they are sprites
again and it is working in unity

Ignoring ProjectSettings

Changes to SpriterDotNetBehaviour and Unity Project Setup

Moved the Spriter Lib output into the SpriterDotNet subfolder of the
Assets folder
Changed the members of SpriterDotNetBehaviour to use camelCasing since
they are not properties.
Changed the SpriterDotNetBehaviour to depend on standard inputs.
Made an animatorSpeed member in SpriterDotNetBehaviour so that it drives
the speed of the real animator and not the other way around. This allows
the user to see the speed in the inpector as well as set it in editor
mode and play mode directly.
Changed NextAnimation to ChangeAnimation so you can go back and forth.
Removed the ProjectSettings from being managed code for reasons of
cross-version compatibility and eventual deployment as a unitypackage.

Fixed reversing animation, resized controls text

Adding back SpriterDotNet dll into the Libs directory

Changed the public fields in the SpriterDotNetBehaviour to use Pascal casing again, and made the single line ifs on the same line.
Made the scene and the prefab use text mode instead of binary mode.

File cleanup

Removed the meta file for the pdb in Libs.