Changes to SpriterDotNetBehaviour and Unity Project Setup
Moved the Spriter Lib output into the SpriterDotNet subfolder of the
Assets folder
Changed the members of SpriterDotNetBehaviour to use camelCasing since
they are not properties.
Changed the SpriterDotNetBehaviour to depend on standard inputs.
Made an animatorSpeed member in SpriterDotNetBehaviour so that it drives
the speed of the real animator and not the other way around. This allows
the user to see the speed in the inpector as well as set it in editor
mode and play mode directly.
Changed NextAnimation to ChangeAnimation so you can go back and forth.
Removed the ProjectSettings from being managed code for reasons of
cross-version compatibility and eventual deployment as a unitypackage.