Fixed deserialization error on fresh project load
Moved everything currently in the awake method to be unity editor only.
AssetDatabase is in the UnityEditor nameSpace which won't be available
at run, so Spriter.Parse will fail.
Also, I was running into issues where it was trying to parse the
SpriterData before it had been properly deserialized, so I have added
the ISerializationCallbackReceiver so that the deserializationComplete
flag can be set and the application can get the green light to parse.